Temporary workaround for button cache issues
Workaround to remove incorrect entries in the duplicate results due to
button cache issues. Specifically, after a duplicate building pair has
been labeled using the button in the Folium popup, the main buttons in
the show_candidate.html are not refreshed and clicking on them sends a
request to the /store-label endpoint using the old id (i.e. of the last
building labeled using the Folium popup). As a quick fix, we remove the
duplicate entries before persisting the results, however, the building
is still shown twice in the browser, likely confusing the users.