Only show candidate pairs during bulk labeling
During neighborhood-wise bulk labeling (using /show-neighborhood), only
show candidate pairs and no other nearby buildings.

Also displaying nearby buildings can help to understand the context and
identify issues such as a structural shift of all buildings in a dataset.

However, it introduces ambiguity because it is unclear which buildings
are candidate pairs (with no edge between them) and which are just
displayed for context. Consequently, it is unclear which buildings
should be labeled. Therefore, now only candidate pair buildings are
displayed.