Gracefully handle edge case where building pair is too distant
Existing buildings are only displayed within 150m around the new
candidate building. If the distance between the pair is bigger - due to
not specifying a max distance when creating the labeling dataset - the
respective existing building of the candidate pair will not be
displayed. Treat such cases gracefully without raising an error.