Support multi-user labeling
Add support for multiple users labeling the same building pairs to
cross-validate previous annotations and achieve censensus for
inconsistently labeled buildings.

Technically, this is done by adding a checkbox to the /index page that
sets a Flask session flag when the "Get Started!" button is clicked,
which is respected by all subsequent /show-pair and /show-neighborhood
endpoint calls. To unset the flag, navigate back to /index and click
"Get Started!" again without checking the box.

Note:
* When ticked, /show-neighborhood shows only neighborhoods labeled
  exactly once.
* When ticked, /show-pair shows pairs labeled exactly once, as well as
  pairs labeled multiple times but ambigiously (i.e. an equal number of
  "Match" and "No Match" labels.