Use factory pattern to create Flask app
Instead of defining the Flask app as a global variable, we use an
application factory with blueprints, which allows us to more cleanly
perform setup operations like initializing the State class or in the
future initialize a db, and also makes testing and configuration
management easier.

Further, do not hard-code filepaths for the LabelingDataset and the
State, but explicitly pass them when loading or initializing the object.