Update how missing matches are added for /show-neighborhood
Instead of using the Folium Draw plugin to manually draw lines to
indicate matches, new matches can be added by subsequently clicking on
an existing and a new building. A line is automatically added to provide
visual feedback to the user.

Further improvements: Instead of sending the geometries of drawn lines
to the backend and trying to identify which buildings were implied, we
identify the id of the clicked building in the frontend and send only id
pairs to the backend.

Note: Candidate pair labels need to be updated first with removed
matches, and then with added matches, to account for the case where
users accidentally delete an existing match and then add it back.