Only add match if it does not exists
Before adding a new match in /show-neighborhood, check if the match
already exists to avoid duplicates and other ambiguities.

Also, check if the added match has been (accidentally) removed before,
in which case it is only removed from the "removedMatches" and not added
as a new match. This allows users to undo changes, without needing to
deal with contradictory added and removed matches in the backend.