Add additional labeling mode
Previously, only labeling of unlabeled buildings and cross-checking of
labeled buildings was supported. Here we add an option to label
everything again from the beginning (or wherever the user left off).

The default labeling mode remains unchanged: resume labeling of
unlabeled buildings.

Technically, the boolean cross_validate flag stored in the Flask session
was replaced with a categorical label_mode variable. The label_mode can
be selected using a radio button on the home page.

It is ensured that username and label_mode are set to default values if
not explicitly set. The results of labeling in incognito mode are saved
under the username 'unknown'.