Retain only max-overlap pairs per dataset to reduce false candidates
Previously, all overlapping buildings between the existing and new
dataset were included as candidate pairs. This resulted in a large
number of candidates, since due to positional inaccuracies and overall
misalignment of the two datasets, many buildings also slightly overlap
with the neighbors of their respective matches.

However, to identify all matches between two datasets (including n:m
relationships), it is sufficient to determine whether the pair with the
largest overlap is a match - as long as for each building from both
datasets, the pair with the largest overlap is retained. This
symmetrical approach ensures that the selected 1:1 candidate pairs can
be aggregated to n:m pairs.