Simplify storage of labeling results by always writing CSV
Previously, labeled pairs were stored as a pickled DataFrame after each
annotation and written to CSV only at the end of the session. This
distinction is unnecessary (at the moment), as persisting 10,000 labeled
pairs to CSV takes only ~15ms.

Now, results are always stored as a CSV file. This also ensures progress
is written to a mounted volume in dockerized deployments, not the
temporary file system of the container, avoiding accidental data loss
during rebuilds or restarts.